+++ rust/cli/src/tui/mod.rs
@@ -1,1 +1,3 @@
 pub mod collection_browser;
+pub mod style;
+pub mod widgets;
